# Use faster multi-contains in `rlike` regex rewrite