Closes #11729 

We have `GpuMultipleContains` to check if the input column contains any of the strings in a list, for [regex rewrite](https://github.com/NVIDIA/spark-rapids/pull/10977). After that, we made a performance improvement for multiple string contains, which provides a similar function and runs fast.

This PR leverages methods in `GpuMultiContains` to `GpuMultipleContains`, and renames it to `GpuContainsAny`. So regex rewrite can also get benefit from the multiple string contains performance improvement.

Performance results on a 1000000 string dataset of 5000~10000 length, averaged three times:

| Search List Length | Pattern Length | Branch 25.02 | This PR (for each ors) | This PR (AST)  | Speedup      |
|--------------------|----------------|---------------|---------|--------------|---|
| 100                | 5              | 5755 ms      | 4064 ms| 3211 ms |79.2%          |
| 100                | 100            | 5537 ms      | 4272 ms| 3283 ms |68.7%        |
| 5                  | 100           | 3514 ms        | 3516 ms  | 2696 ms  | 30.3%           |
| 2                 | 10              | 3597 ms       | 3460ms | 2746 ms |31.0%       |

<!--

Thank you for contributing to RAPIDS Accelerator for Apache Spark!

Here are some guidelines to help the review process go smoothly.

1. Please write a description in this text box of the changes that are being
   made.

2. Please ensure that you have written units tests for the changes made/features
   added.

3. If you are closing an issue please use one of the automatic closing words as
   noted here: https://help.github.com/articles/closing-issues-using-keywords/

4. If your pull request is not ready for review but you want to make use of the
   continuous integration testing facilities please label it with `[WIP]`.

5. If your pull request is ready to be reviewed without requiring additional
   work on top of it, then remove the `[WIP]` label (if present).

6. Once all work has been done and review has taken place please do not add
   features or make changes out of the scope of those requested by the reviewer
   (doing this just add delays as already reviewed code ends up having to be
   re-reviewed/it is hard to tell what is new etc!). Further, please avoid
   rebasing your branch during the review process, as this causes the context
   of any comments made by reviewers to be lost. If conflicts occur during
   review then they should be resolved by merging into the branch used for
   making the pull request.

Many thanks in advance for your cooperation!

-->