@thirtiseven 2024-12-06
> Could you redo your speedup numbers? Mixing % and x speedup factors is confusing. Also the 2.31x number appears to be wrong. That or the times are wrong because
> 
> 6,399 ms / 3,765 ms is 1.70x speedup (70%) not 2.31x (131%).

Thanks, updated in the description.