@thirtiseven 2024-12-13
Profiled a 100 pattens with 5 length case:
Before this pr:
<img width="1040" alt="Screenshot 2024-12-13 at 13 58 49" src="https://github.com/user-attachments/assets/1c28c1a2-95e2-4723-9df5-b3dbbd8ef20b" />

After this pr:
<img width="1052" alt="Screenshot 2024-12-13 at 13 59 56" src="https://github.com/user-attachments/assets/f482cd8f-aa64-4a21-9a58-1c4fe44a747f" />

There is a 3~5x speedup in the operator level, the `contains` are quite fast and the bottleneck now is the multiple `or`s. We can write a simple custom kernel to do multiple `or` to make it much faster (maybe as a followup).