@revans2 2024-12-13
> There is a 3~5x speedup in the operator level, the `contains` are quite fast and the bottleneck now is the multiple `or`s. We can write a simple custom kernel to do multiple `or` to make it much faster (maybe as a followup).


Could you try using AST to do the ORs?  An alternative would be to have a separate multi-contains that does the or automatically.