@thirtiseven 2024-12-16
> Could you try using AST to do the ORs? An alternative would be to have a separate multi-contains that does the or automatically.


Thanks @revans2, tried AST and it looks to work with another 1x speed up in the operator level:

<img width="928" alt="Screenshot 2024-12-16 at 17 52 54" src="https://github.com/user-attachments/assets/d4fc97f0-745d-4c35-84fb-4c9ff0a46c38" />

Updated the performance results too.

Not sure if I coded it right, please take another look.